expose/requirements.txt
python-docx==0.8.11
PyPDF2==3.0.1
pandas==2.0.3
openpyxl==3.1.2
python-hwp==0.1.0
reportlab==4.0.4
sentence-transformers==2.2.2
transformers==4.30.2
torch==2.0.1
scikit-learn==1.2.2
numpy==1.24.3
pytesseract==0.3.10
Pillow==9.5.0 